refactor: rename package of ad-hoc tools schema connector
