+++ connectors/agentic-ai/agentic-ai-connectors/src/main/java/io/camunda/connector/agenticai/adhoctoolsschema/CachingAdHocToolsSchemaResolver.java
@@ -4,7 +4,7 @@
  * See the License.txt file for more information. You may not use this file
  * except in compliance with the proprietary license.
  */
-package io.camunda.connector.agenticai.schema;
+package io.camunda.connector.agenticai.adhoctoolsschema;
 
 import com.github.benmanes.caffeine.cache.Caffeine;
 import com.github.benmanes.caffeine.cache.LoadingCache;
