+++ connectors/agentic-ai/agentic-ai-connectors/src/main/java/io/camunda/connector/agenticai/util/ObjectMapperConstants.java
@@ -4,7 +4,7 @@
  * See the License.txt file for more information. You may not use this file
  * except in compliance with the proprietary license.
  */
-package io.camunda.connector.agenticai.mapping;
+package io.camunda.connector.agenticai.util;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import java.util.Map;
